Moved messageText state, query for adding new messages to firebase, and JSX form into its own component, MessageForm.
