+++ client/src/views/MessagingContainer.js
@@ -1,11 +1,10 @@
 import React, { useState, useEffect } from "react";
-import { connect } from "react-redux";
 import MessageList from "./MessageList";
+import MessageForm from "./MessageForm";
 import firebase from "firebase";
 import "./styles/Messaging.scss";
 
-function Messaging({ user }) {
-  const [messageText, setMessageText] = useState("");
+function MessagingContainer() {
   const [messageCollection, setMessageCollection] = useState([]);
   useEffect(() => {
     console.log("loading messages");
@@ -44,27 +43,6 @@ function Messaging({ user }) {
       });
     });
   };
-
-  const submitMessage = e => {
-    e.preventDefault();
-    const { email, photoUrl } = user;
-    // Add a new message entry to the Firebase database.
-    return firebase
-      .firestore()
-      .collection("messages")
-      .add({
-        email: email,
-        text: messageText,
-        profileImageUrl: photoUrl,
-        timestamp: firebase.firestore.FieldValue.serverTimestamp()
-      })
-      .then(() => {
-        setMessageText("");
-      })
-      .catch(err => {
-        console.error("Error writing new message to Firebase Database", err);
-      });
-  };
   return (
     <main className="mdl-layout__content mdl-color--grey-100">
       <div
@@ -77,29 +55,7 @@ function Messaging({ user }) {
         >
           <div className="mdl-card__supporting-text mdl-color-text--grey-600">
             <MessageList messages={messageCollection} />
-            <form id="message-form" onSubmit={submitMessage}>
-              <div className="mdl-textfield mdl-js-textfield mdl-textfield--floating-label">
-                <input
-                  className="mdl-textfield__input"
-                  type="text"
-                  id="message"
-                  autoComplete="off"
-                  value={messageText}
-                  onChange={e => setMessageText(e.target.value)}
-                />
-                <label className="mdl-textfield__label" htmlFor="message">
-                  Message...
-                </label>
-              </div>
-              <button
-                id="submit"
-                disabled={messageText.length === 0}
-                type="submit"
-                className="mdl-button mdl-js-button mdl-button--raised mdl-js-ripple-effect"
-              >
-                Send
-              </button>
-            </form>
+            <MessageForm />
             <form id="image-form">
               <input
                 id="mediaCapture"
@@ -122,11 +78,4 @@ function Messaging({ user }) {
   );
 }
 
-const mapStateToProps = state => ({
-  user: state.authReducer.user
-});
-
-export default connect(
-  mapStateToProps,
-  {}
-)(Messaging);
+export default MessagingContainer;
